+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Naive implementation of Yaz0 compression.
+
 ## [0.2.0] - 2020-09-13
 
 ### Added